# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MissionOut API",
    "description": "MissionOut API contract for dispatcher, team admin, and responder clients. The exported OpenAPI document in contracts/openapi.json is the cross-stack source of truth.",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/email-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request Email Sign-In Code",
        "description": "Starts email-based sign-in by asking the backend to send a one-time verification code to the supplied email address. The code is entered directly into the MissionOut client for the requested surface.",
        "operationId": "request_email_code_auth_email_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailCodeSentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/email-code/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Email Sign-In Code",
        "description": "Completes email-code sign-in by exchanging a one-time emailed code for the authenticated MissionOut user payload.",
        "operationId": "verify_email_code_auth_email_code_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Google Auth",
        "operationId": "google_auth_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices/web-push": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register Web Push Subscription",
        "operationId": "register_web_push_subscription_devices_web_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebPushSubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebPushSubscriptionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Delete Web Push Subscription",
        "operationId": "delete_web_push_subscription_devices_web_push_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebPushSubscriptionDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthTeamMembershipRead": {
        "properties": {
          "team_public_id": {
            "type": "string",
            "title": "Team Public Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "team_public_id",
          "team_name",
          "roles"
        ],
        "title": "AuthTeamMembershipRead"
      },
      "AuthUserRead": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "initials": {
            "type": "string",
            "title": "Initials"
          },
          "global_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Global Permissions"
          },
          "team_memberships": {
            "items": {
              "$ref": "#/components/schemas/AuthTeamMembershipRead"
            },
            "type": "array",
            "title": "Team Memberships"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "name",
          "initials",
          "global_permissions",
          "team_memberships",
          "email"
        ],
        "title": "AuthUserRead"
      },
...
      "EmailCodeRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "requested_client": {
            "type": "string",
            "enum": [
              "responder",
              "dispatcher",
              "team_admin"
            ],
            "title": "Requested Client"
          }
        },
        "type": "object",
        "required": [
          "email",
          "requested_client"
        ],
        "title": "EmailCodeRequest"
      },
      "EmailCodeSentRead": {
        "properties": {
          "delivery": {
            "type": "string",
            "const": "email_code",
            "title": "Delivery",
            "default": "email_code"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "expires_in_minutes": {
            "type": "integer",
            "title": "Expires In Minutes"
          },
          "code_length": {
            "type": "integer",
            "title": "Code Length",
            "default": 6
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "email",
          "expires_in_minutes",
          "message"
        ],
        "title": "EmailCodeSentRead"
      },
      "EmailCodeVerifyRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "title": "EmailCodeVerifyRequest"
      },
      "GoogleAuthRequest": {
        "properties": {
          "id_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Token"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "requested_client": {
            "type": "string",
            "enum": [
              "responder",
              "dispatcher",
              "team_admin"
            ],
            "title": "Requested Client"
          }
        },
        "type": "object",
        "required": [
          "requested_client"
        ],
        "title": "GoogleAuthRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebPushKeys": {
        "properties": {
          "p256dh": {
            "type": "string",
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "WebPushKeys"
      },
...
      "WebPushSubscriptionCreate": {
        "properties": {
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "keys": {
            "$ref": "#/components/schemas/WebPushKeys"
          },
          "user_agent": {
            "type": "string",
            "title": "User Agent",
            "default": ""
          },
          "client": {
            "type": "string",
            "enum": [
              "responder",
              "dispatcher",
              "team_admin"
            ],
            "title": "Client"
          },
          "team_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Public Id"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys",
          "client"
        ],
        "title": "WebPushSubscriptionCreate"
      },
      "WebPushSubscriptionDelete": {
        "properties": {
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "endpoint"
        ],
        "title": "WebPushSubscriptionDelete"
      },
      "WebPushSubscriptionRead": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "user_public_id": {
            "type": "string",
            "title": "User Public Id"
          },
          "team_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Public Id"
          },
          "platform": {
            "type": "string",
            "const": "web",
            "title": "Platform",
            "default": "web"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "client": {
            "type": "string",
            "title": "Client"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "user_public_id",
          "team_public_id",
          "endpoint",
          "client",
          "last_seen",
          "is_active"
        ],
        "title": "WebPushSubscriptionRead"
      }
    }
  }
}